change types to i32 to avoid overflow

+++ src/game.rs
@@ -13,10 +13,10 @@ use rand::{thread_rng, Rng};
 pub struct Game {
     board: Board,
     next_piece: Piece,
-    pos_x: usize,
-    pos_y: usize,
-    next_pos_x: usize,
-    next_pos_y: usize,
+    pos_x: i32,
+    pos_y: i32,
+    next_pos_x: i32,
+    next_pos_y: i32,
     piece_type: PieceType,
     next_piece_type: PieceType,
     rotation: i32,
@@ -37,10 +37,10 @@ impl Game {
 
         let piece           = Piece::new(piece_type, rotation);
         let next_piece      = Piece::new(next_piece_type, next_rotation);
-        let board           = Board::new(piece, m_screen_height as usize);
+        let board           = Board::new(piece, m_screen_height);
 
-        let pos_x           = Board::get_pos_x() as usize;
-        let pos_y           = Piece::get_y_initial_position(&board.piece) as usize;
+        let pos_x           = Board::get_pos_x();
+        let pos_y           = Piece::get_y_initial_position(&board.piece);
 
         let next_pos_x = 10;
         let next_pos_y = 0;
@@ -74,15 +74,15 @@ impl Game {
         std::mem::swap(&mut new_piece, &mut self.next_piece);
         self.board.update_current_piece(new_piece);
 
-        self.pos_x = Board::get_pos_x() as usize;
-        self.pos_y = self.board.piece.get_y_initial_position() as usize;
+        self.pos_x = Board::get_pos_x();
+        self.pos_y = self.board.piece.get_y_initial_position();
         self.next_piece_type = Piece::get_piece(Game::get_rand(1, 7));
         self.next_rotation = Game::get_rand(0, 3);
     }
 
-    pub fn draw_piece(&self, x: usize, y: usize, piece_to_draw: &Piece) { 
+    pub fn draw_piece(&self, x: i32, y: i32, piece_to_draw: &Piece) { 
         // Board::draw_piece(&self.board, x, y, piece_to_draw);
-        self.board.draw_piece(x, y, piece_to_draw);
+        self.board.draw_piece(x, y, piece_to_draw); // STACK 3
     }
 
     pub fn draw_board(&self) {
@@ -94,7 +94,7 @@ impl Game {
         View::clear_screen();
         View::load_background();
         self.draw_board();
-        self.draw_piece(self.pos_x, self.pos_y, &self.board.piece);
+        self.draw_piece(self.pos_x, self.pos_y, &self.board.piece); // STACK 4
         self.draw_piece(self.pos_x, self.pos_y, &self.get_next_piece());
 
         View::update_screen();
